fix: separate and modify components
