+++ peer_web_frontend/src/app/find-account/page.tsx
@@ -3,8 +3,7 @@
 import { Typography, Container, IconButton, Box } from '@mui/material'
 import SendEmailForm from './panel/SendEmailForm'
 import useMedia from '@/hook/useMedia'
-import { NavigateBefore } from '@mui/icons-material'
-import { useRouter } from 'next/navigation'
+import BackIconButton from '@/components/BackIconButton'
 
 const PCBase = {
   display: 'flex',
@@ -44,7 +43,6 @@ const MobileBox = {
 
 const FindAccount = () => {
   const { isPc } = useMedia()
-  const router = useRouter()
 
   return (
     <Container sx={isPc ? PCBase : MobileBase}>
@@ -61,12 +59,7 @@ const FindAccount = () => {
               marginBottom: '16px',
             }}
           >
-            <IconButton
-              sx={{ width: '40px', height: '40px' }}
-              onClick={() => router.back()}
-            >
-              <NavigateBefore />
-            </IconButton>
+            <BackIconButton />
             <Typography>비밀번호 찾기</Typography>
             <IconButton
               sx={{ width: '40px', height: '40px' }}
